Move formatTime from core:ui to core:common
The audiolog recorder (in core:playback) needs to format positions for its
labels, but core:playback must not depend on the Compose UI module core:ui.
formatTime is a pure helper, so moving it to the lower-level core:common
module lets both core:ui and core:playback reuse a single implementation.
